// packages/apps-wake/src/utils/graphql/storefront.graphql.gen.ts
export type BuyListQuery = {
  buyList?: {
    mainVariant?: boolean | null;
    productName?: string | null;
    productId?: any | null;
    alias?: string | null;
    collection?: string | null;
    kit: boolean;
    numberOfVotes?: number | null;
    available?: boolean | null;
    averageRating?: number | null;
    condition?: string | null;
    createdAt?: any | null;
    ean?: string | null;
    id?: string | null;
    minimumOrderQuantity?: number | null;
    productVariantId?: any | null;
    sku?: string | null;
    stock?: any | null;
    variantName?: string | null;
    parallelOptions?: Array<string | null> | null;
    urlVideo?: string | null;
    buyListId: number;
    attributes?: Array<{
      name?: string | null;
      type?: string | null;
      value?: string | null;
      attributeId: any;
      displayType?: string | null;
      id?: string | null;
    } | null> | null;
    productCategories?: Array<{
      id: number;
      name?: string | null;
      url?: string | null;
      hierarchy?: string | null;
      main: boolean;
      googleCategories?: string | null;
    } | null> | null;
    informations?: Array<{
      title?: string | null;
      value?: string | null;
      type?: string | null;
      id: any;
    } | null> | null;
    breadcrumbs?: Array<{ text?: string | null; link?: string | null } | null> | null;
    images?: Array<{ url?: string | null; fileName?: string | null; print: boolean } | null> | null;
    prices?: {
      discountPercentage: any;
      discounted: boolean;
      listPrice?: any | null;
      multiplicationFactor: number;
      price: any;
      bestInstallment?: {
        discount: boolean;
        displayName?: string | null;
        fees: boolean;
        name?: string | null;
        number: number;
        value: any;
      } | null;
      installmentPlans?: Array<{
        displayName?: string | null;
        name?: string | null;
        installments?: Array<{
          discount: boolean;
          fees: boolean;
          number: number;
          value: any;
        } | null> | null;
      } | null> | null;
      priceTables?: Array<{
        discountPercentage: any;
        id: any;
        listPrice?: any | null;
        price: any;
      } | null> | null;
      wholesalePrices?: Array<{ price: any; quantity: number } | null> | null;
    } | null;
    productBrand?: {
      fullUrlLogo?: string | null;
      logoUrl?: string | null;
      name?: string | null;
      alias?: string | null;
    } | null;
    seller?: { name?: string | null } | null;
    seo?: Array<{
      name?: string | null;
      scheme?: string | null;
      type?: string | null;
      httpEquiv?: string | null;
      content?: string | null;
    } | null> | null;
    reviews?: Array<{
      rating: number;
      review?: string | null;
      reviewDate: any;
      email?: string | null;
      customer?: string | null;
    } | null> | null;
    similarProducts?: Array<{
      alias?: string | null;
      image?: string | null;
      imageUrl?: string | null;
      name?: string | null;
    } | null> | null;
    attributeSelections?: {
      canBeMatrix: boolean;
      selections?: Array<{
        attributeId: any;
        displayType?: string | null;
        name?: string | null;
        varyByParent: boolean;
        values?: Array<{
          alias?: string | null;
          available: boolean;
          value?: string | null;
          selected: boolean;
          printUrl?: string | null;
        } | null> | null;
      } | null> | null;
      matrix?: {
        column?: {
          displayType?: string | null;
          name?: string | null;
          values?: Array<{ value?: string | null } | null> | null;
        } | null;
        data?: Array<Array<{
          available: boolean;
          productVariantId: any;
          stock: any;
        } | null> | null> | null;
        row?: {
          displayType?: string | null;
          name?: string | null;
          values?: Array<{ value?: string | null; printUrl?: string | null } | null> | null;
        } | null;
      } | null;
    } | null;
    buyTogether?: Array<{ productId?: any | null } | null> | null;
    promotions?: Array<{
      content?: string | null;
      disclosureType?: string | null;
      id: any;
      fullStampUrl?: string | null;
      stamp?: string | null;
      title?: string | null;
    } | null> | null;
    buyListProducts?: Array<{
      productId: any;
      quantity: number;
      includeSameParent: boolean;
    } | null> | null;
  } | null;
};